+++ src/shared/chat/emoji-picker-content.js
@@ -2,6 +2,7 @@ import { CustomElement } from 'shared/components/element.js';
 import { _converse, converse, api } from "@converse/headless/core";
 import { html } from "lit";
 import { tpl_all_emojis, tpl_search_results } from "./templates/emoji-picker.js";
+import { getTonedEmojis } from './utils.js';
 
 const { sizzle } = converse.env;
 
@@ -90,7 +91,7 @@ export default class EmojiPickerContent extends CustomElement {
               return true;
           }
       } else {
-          if (this.current_skintone && converse.emojis.toned.includes(shortname)) {
+          if (this.current_skintone && getTonedEmojis().includes(shortname)) {
               return true;
           }
       }
